pyproject.toml
[tool.poetry]
name = "dagbot"
version = "2.9.1"
description = "Fun Meme Discord bot"
authors = ["Daggy1234"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.1"
aiohttp = "3.7.3"
"discord.py" = {git = "https://github.com/rapptz/discord.py", branch="performance"}
art = "^5.1"
sr-api = "^1.1.3"
async-cse = "^0.3.0"
async-cleverbot = "^0.2.2"
async-timeout = "^3.0.1"
jishaku = {git="https://github.com/Gorialis/jishaku/"}
tabulate = "^0.8.7"
psutil = "^5.8.0"
humanize = "^3.2.0"
asyncdagpi = {git = "https://github.com/daggy1234/asyncdagpi"}
owotext = "^1.1.1"
asyncpg = "^0.21.0"
beautifulsoup4 = "4.9.3"
zalgoify = "^1.1"
PyYAML = "^5.3.1"
sentry-sdk = "^1.0.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
cryptography = "3.2"
vaporwavely = "^1.0.1"
PyDictionary = "^2.0.1"
validator-collection = "^1.5.0"
matplotlib = "^3.3.3"
RandomWords = "^0.3.0"
Pillow = "8.1.1"
numpy = "1.19.3"
polaroid = "^0.3.2"


[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
autopep8 = "^1.5.4"
yapf = "^0.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
